erms-1357

+++ app/grails-app/controllers/com/k_int/kbplus/SubscriptionController.groovy
@@ -308,7 +308,7 @@ class SubscriptionController extends AbstractDebugController {
     def delete() {
         def result = setResultGenericsAndCheckAccess(AccessService.CHECK_EDIT)
 
-        if (params.process) {
+        if (params.process  && result.editable) {
             result.result = deletionService.deleteSubscription(result.subscription, false)
         }
         else {
